requirements.txt
pillow==12.0.0
numpy==2.3.5
scikit-image==0.25.2
scikit-learn==1.4.2
pandas==2.2.3
joblib==1.4.2
psutil==7.1.3
albumentations==1.4.20
opencv-python==4.10.0.84
matplotlib==3.9.2
seaborn==0.13.2
